requirements.txt
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
mpmath==1.3.0
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
pyglet==2.1.0
Pygments==2.18.0
pytest==8.3.4
rich==13.9.4
scipy==1.14.1
sympy==1.13.3
